move indexing tools into utils._indexing
