+++ sklearn/compose/_column_transformer.py
@@ -18,8 +18,9 @@ from scipy import sparse
 from ..base import TransformerMixin, _fit_context, clone
 from ..pipeline import _fit_transform_one, _name_estimators, _transform_one
 from ..preprocessing import FunctionTransformer
-from ..utils import Bunch, _get_column_indices, _safe_indexing
+from ..utils import Bunch, _safe_indexing
 from ..utils._estimator_html_repr import _VisualBlock
+from ..utils._indexing import _get_column_indices
 from ..utils._metadata_requests import METHODS
 from ..utils._param_validation import HasMethods, Hidden, Interval, StrOptions
 from ..utils._set_output import (
